+++ boto3_refresh_session/__init__.py
@@ -10,9 +10,11 @@ from .methods.custom import *
 from .methods.iot import *
 from .methods.sts import *
 from .session import *
-from .utils import config
+from .utils import cache, config
+from .utils.cache import *
 from .utils.config import *
 
+__all__.extend(cache.__all__)
 __all__.extend(config.__all__)
 __all__.extend(session.__all__)
 __all__.extend(exceptions.__all__)